add bitnum.h
Since the syscall set that we care about only has 57 elements. I
use a unsigned long (64 bits) variable to indicates which subset
of the syscalls we want to forward. E.g. the bit number of sys_brk
is 1. Thus, if we want to forward sys_brk, we set the 1st bit of
the variable to 1.